requirements.txt
langchain==0.3.19
openai==1.64.0
streamlit==1.42.0
langchain-openai==0.3.6
langchain-community==0.3.17
tiktoken==0.9.0
pillow==11.1.0
pypdf==5.3.0
pymongo==4.11.1
tabulate==0.9.0
azure-identity==1.20.0
azure-search-documents==11.5.2
faiss-cpu==1.10.0
tavily-python==0.5.1
langgraph==0.2.74